src/test/java/dev/drf/leetcode/number/of/islands/NumberOfIslandsProblemTest.java
package dev.drf.leetcode.number.of.islands;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class NumberOfIslandsProblemTest {
    private final NumberOfIslandsProblem problem = new NumberOfIslandsProblem();

    @Test
    void example1() {
        /*
        Input: grid = [
          ["1","1","1","1","0"],
          ["1","1","0","1","0"],
          ["1","1","0","0","0"],
          ["0","0","0","0","0"]
        ]
        Output: 1
         */
        char[][] grid = new char[][]{
                {'1', '1', '1', '1', '0'},
                {'1', '1', '0', '1', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '0', '0', '0'}
        };

        var count = problem.numIslands(grid);

        assertEquals(1, count);
    }

    @Test
    void example2() {
        /*
        Input: grid = [
          ["1","1","0","0","0"],
          ["1","1","0","0","0"],
          ["0","0","1","0","0"],
          ["0","0","0","1","1"]
        ]
        Output: 3
         */
        char[][] grid = new char[][]{
                {'1', '1', '0', '0', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '1', '0', '0'},
                {'0', '0', '0', '1', '1'}
        };

        var count = problem.numIslands(grid);

        assertEquals(3, count);
    }
}
